cristhian: refactorization following more the SOLID principles

+++ src/main/java/com/haiilo/kata/backend/model/dto/CartDto.java
@@ -2,13 +2,14 @@ package com.haiilo.kata.backend.model.dto;
 
 import com.haiilo.kata.backend.model.enums.CartStatus;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CartDto(
         Long id,
         CartStatus cartStatus,
         List<CartItemDto> items,
-        Double total,
+        BigDecimal total,
         String currency
 ) {
 }
