+++ src/main/java/com/haiilo/kata/backend/model/dto/CartItemDto.java
@@ -2,12 +2,14 @@ package com.haiilo.kata.backend.model.dto;
 
 import com.haiilo.kata.backend.model.enums.Status;
 
+import java.math.BigDecimal;
+
 public record CartItemDto(
         Long id,
         Long cartId,
         Long productId,
         String productName,
-        Double unitPrice,
+        BigDecimal unitPrice,
         String currency,
         Integer quantity,
         Status status
