+++ src/main/java/com/haiilo/kata/backend/model/dto/ReceiptDto.java
@@ -2,14 +2,15 @@ package com.haiilo.kata.backend.model.dto;
 
 import com.haiilo.kata.backend.model.json.TransactionDetail;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ReceiptDto(
         Long id,
         Long cartId,
-        Double subTotal,
-        Double discount,
-        Double total,
+        BigDecimal subTotal,
+        BigDecimal discount,
+        BigDecimal total,
         String currency,
         List<TransactionDetail> transactionDetails
 ) {
