+++ src/main/java/com/haiilo/kata/backend/model/json/AppliedOffer.java
@@ -1,6 +1,8 @@
 package com.haiilo.kata.backend.model.json;
 
+import java.math.BigDecimal;
+
 public record AppliedOffer(
     String description,
-    Double discountAmount
+    BigDecimal discountAmount
 ) {}
